Move location route
